update in delete product

+++ backend/controllers/products.controller.js
@@ -50,11 +50,15 @@ export const deleteProduct = async(req,res)=> {
     const {id} = req.params
     console.log("id is",id)
 
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({success: false, message: "Invalid Product ID"})
+    }
+
     try {
         await Product.findByIdAndDelete(id)
         res.status(200).json({success: true, message:"Product deleted"})
     } catch(error){
         console.log("error - ", error);
-        res.status(400).json({success:false, message: "Product not found!"})
+        res.status(500).json({success:false, message: "Server Error"})
     }
 }
